analytics/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
openpyxl==3.1.5
reportlab==4.2.5
pandas==2.2.3
numpy==2.2.1
